backend/packages/gizmo-agent/gizmo_agent/ingest.py
import os
from dotenv import load_dotenv

from agent_executor.upload import IngestRunnable
from langchain.embeddings import OpenAIEmbeddings
from langchain.schema.runnable import ConfigurableField
from langchain.text_splitter import RecursiveCharacterTextSplitter
from langchain.vectorstores.chroma import Chroma

load_dotenv()

index_schema = {
    "tag": [{"name": "namespace"}],
}
vstore = Chroma("opengpts", OpenAIEmbeddings(), "backend/vstore")


ingest_runnable = IngestRunnable(
    text_splitter=RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=200),
    vectorstore=vstore,
).configurable_fields(
    assistant_id=ConfigurableField(
        id="assistant_id",
        annotation=str,
        name="Assistant ID",
    ),
)
