backend/requirements.txt
langchain-cli
-e packages/gizmo-agent
-e packages/agent-executor
langchain>=0.0.331
langserve>=0.0.23
python-dotenv
chromadb